slow af

+++ TibberTask/TibberTask/Models/Execution.cs
@@ -2,7 +2,7 @@
 {
     public class Execution
     {
-        public Execution(int id, DateTime timestamp, int commands, int result, double duration)
+        public Execution(int id, DateTime timestamp, int commands, long result, double duration)
         {
             Id = id;
             Timestamp = timestamp;
@@ -13,7 +13,7 @@
         public int Id { get; set; }
         public DateTime Timestamp { get; set; }
         public int Commands { get; set; }
-        public int Result { get; set; }
+        public long Result { get; set; }
         public double Duration { get; set; }
